requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
html2text>=2024.2.26
lxml>=5.1.0
pandas>=2.2.0
pdfkit>=1.0.0
jinja2
python-dotenv